requirements.txt
django
win10toast
selenium
msedge.selenium_tools